refactor: 동일광고 유사율 리스트 조회 info, mapper 수정
- 사용 객체 타입 변경에 따른 변경

+++ src/main/java/com/fastcampus/befinal/domain/info/SameAdInfo.java
@@ -2,6 +2,7 @@ package com.fastcampus.befinal.domain.info;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -23,7 +24,7 @@ public class SameAdInfo {
         String advertiser,
         String category,
         LocalDateTime postDateTime,
-        Double similarity,
+        BigDecimal similarity,
         Integer sameSentenceCount
     ) {}
